Addition of tests, Feature Selection and DMR identification
The PCs are inspected for batch effects and other influences, PC2 is free of batches such as Biomaterial provider, so it's used to select the genes to put under scrutiny for DMR selection. After selecting the genes with highest loading scores p.tests are run to identify DMRs. The ones with too low "statistics" values are discarded and the remaining ones are inspected for multiple comparison correction. Among the selected ones the top 20 are extracted for integrative analysis and logistic regression.